Move test-plan projection out of finding boundary
