+++ apps/server/tests/adapters/pdf/test_summary_view.py
@@ -18,8 +18,8 @@ from vibesensor.shared.boundaries.analysis_summary_projection import project_ana
 from vibesensor.shared.boundaries.diagnostic_case import (
     test_run_from_summary as _test_run_from_summary,
 )
-from vibesensor.shared.boundaries.finding import step_payloads_from_plan
 from vibesensor.shared.boundaries.report_interpretation import resolve_report_origin
+from vibesensor.shared.boundaries.test_plan_projection import step_payloads_from_plan
 from vibesensor.shared.boundaries.vibration_origin import origin_payload_from_finding
